projects/speech_to_text/requirements.txt
openai-whisper
sounddevice
scipy
